cherrypy/requirements.txt
CherryPy
PySensors
